Fully resolve clink path when multiple symlinks are involved
This patch allows clink to correctly infer it's real path
when it's invoked as a symlink to a symlink. It correctly
handles both relative and absolute symlinks.

This is helpful in the following scenario:
lrwxrwxrwx 1 jp jp        5 Apr  1 14:26 aws -> clink
lrwxrwxrwx 1 jp jp       31 Apr  1 14:24 clink -> /home/jp/git/clink/clink